+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `castToBigInt` and `convertBigIntForDatasource` to `Datasource` class
 * Add `LocalDateTime`, `LocalDate` and `LocalTime` fields
 * Make `TaskSchedules#startFromMenu()` return the started instance
+* Add `Field.Decimal` and `Field.FixedDecimal` fields
 
 ## 1.3.19 (2023-10-18)